Performance optimization: 50-98% speedup across key operations

+++ benches/asymmetric_convergence.rs
@@ -27,17 +27,17 @@ use computable::Bounds;
 use computable::{Binary, Computable, pi};
 
 fn sqrt_2() -> Computable {
-    Computable::constant(Binary::new(BigInt::from(2_i64), BigInt::from(0_i64)))
+    Computable::constant(Binary::new(BigInt::from(2_i64), 0_i64))
         .nth_root(NonZeroU32::new(2_u32).unwrap())
 }
 
 fn cbrt_3() -> Computable {
-    Computable::constant(Binary::new(BigInt::from(3_i64), BigInt::from(0_i64)))
+    Computable::constant(Binary::new(BigInt::from(3_i64), 0_i64))
         .nth_root(NonZeroU32::new(3_u32).unwrap())
 }
 
 fn constant(value: i64) -> Computable {
-    Computable::constant(Binary::new(BigInt::from(value), BigInt::from(0_i64)))
+    Computable::constant(Binary::new(BigInt::from(value), 0_i64))
 }
 
 // --- mixed (asymmetric) ---
